refactor(nostr_sdk): keep the query outcome tracker off the public API
QueryOutcomeTracker is RelayPool's internal bookkeeping, so it moves to lib/src under @internal and only QueryOutcome stays importable.
